# platform-api/plugins/eventgateway/openapi.yaml
openapi: "3.0.3"
info:
  title: Event Gateway Plugin API
  version: "0.9"
  description: >
    WebSub and WebBroker API management endpoints provided by the
    eventgateway compile-time plugin.
paths:
...
  /api/v0.9/websub-apis/{apiId}/deployments:
    parameters:
      - name: apiId
        in: path
        required: true
        schema:
          type: string
    post:
      security:
        - OAuth2SecurityEventGateway:
            - ap:websub_api:deployment:create
            - ap:websub_api:deployment:manage
      responses:
        '201':
          description: WebSub API deployment created successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      security:
        - OAuth2SecurityEventGateway:
            - ap:websub_api:deployment:read
            - ap:websub_api:deployment:manage
      responses:
        '200':
          description: WebSub API deployments retrieved successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v0.9/websub-apis/{apiId}/deployments/{deploymentId}:
    parameters:
      - name: apiId
        in: path
        required: true
        schema:
          type: string
      - name: deploymentId
        in: path
        required: true
        schema:
          type: string
    get:
      security:
        - OAuth2SecurityEventGateway:
            - ap:websub_api:deployment:read
            - ap:websub_api:deployment:manage
      responses:
        '200':
          description: WebSub API deployment retrieved successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      security:
        - OAuth2SecurityEventGateway:
            - ap:websub_api:deployment:delete
            - ap:websub_api:deployment:manage
      responses:
        '204':
          description: WebSub API deployment deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v0.9/websub-apis/{apiId}/deployments/{deploymentId}/undeploy:
    parameters:
      - name: apiId
        in: path
        required: true
        schema:
          type: string
      - name: deploymentId
        in: path
        required: true
        schema:
          type: string
    post:
      security:
        - OAuth2SecurityEventGateway:
            - ap:websub_api:deployment:undeploy
            - ap:websub_api:deployment:manage
      responses:
        '200':
          description: WebSub API deployment undeployed successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v0.9/websub-apis/{apiId}/deployments/{deploymentId}/restore:
    parameters:
      - name: apiId
        in: path
        required: true
        schema:
          type: string
      - name: deploymentId
        in: path
        required: true
        schema:
          type: string
    post:
      security:
        - OAuth2SecurityEventGateway:
            - ap:websub_api:deployment:restore
            - ap:websub_api:deployment:manage
      responses:
        '200':
          description: WebSub API deployment restored successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  responses:
    BadRequest:
      description: Bad request
    Unauthorized:
      description: Unauthorized
    NotFound:
      description: Not found
    Conflict:
      description: Conflict
...
    InternalServerError:
      description: Internal server error
...
  securitySchemes:
    OAuth2SecurityEventGateway:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /oauth2/token
          scopes: {}